Service/face_recognition/requirements.txt
face_recognition
numpy
Pillow 